07/15(삭제페이지-게시물삭제(방법2-valid))

+++ Spring/src/main/java/kr/green/spring/service/BoardServiceImp.java
@@ -65,6 +65,7 @@ public class BoardServiceImp implements BoardService {
 	
 	@Override
 	public void deleteBoard(BoardVO bVO) {
+		if(bVO.getNum() <= 0 ) return ;
 		boardDao.deleteBoard(bVO);
 		
 	}
